generated_agents/1750224705337jelcldqfm/general_agent/python_agent/requirements.txt
openai>=1.3.0
google-auth>=2.16.0
google-auth-oauthlib>=0.7.1
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
python-dotenv>=0.19.0
requests>=2.28.0